# docs/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spektors API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/internal/admin/v1/threads": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "Список тредов для инбокса",
        "description": "Последние активные треды с превью последнего сообщения. Опциональный **`app_id`** — только известные id из `config/apps.yaml`.",
        "operationId": "list_threads_inbox_internal_admin_v1_threads_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "app_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по приложению (web_client, web_lite, …)",
              "title": "App Id"
            },
            "description": "Фильтр по приложению (web_client, web_lite, …)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminThreadsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Токен обязателен, но не передан"
          },
          "403": {
            "description": "Неверный Bearer-токен"
          },
          "404": {
            "description": "Неизвестный app_id (если передан фильтр)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/admin/v1/sessions/{session_id}/threads/{thread_id}/reply": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Ответ оператора в чат",
        "description": "Добавляет сообщение `assistant` с `msg_source=human`. Тред должен существовать. Если на сервере задан **`INTERNAL_ADMIN_TOKEN`**, нужен заголовок **`Authorization: Bearer <токен>`** (в Swagger — **HTTP Bearer**). `session_id` и `thread_id` в пути — hex 8–64, без дефисов.",
        "operationId": "operator_reply_to_thread_internal_admin_v1_sessions__session_id__threads__thread_id__reply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 64,
              "pattern": "^[a-fA-F0-9]+$",
              "description": "Идентификатор сессии или треда: hex 8–64 символов (как `uuid.uuid4().hex`), без дефисов",
              "examples": [
                "a1b2c3d4e5f6789012345678abcdef01"
              ],
              "title": "Session Id"
            },
            "description": "Идентификатор сессии или треда: hex 8–64 символов (как `uuid.uuid4().hex`), без дефисов"
          },
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 64,
              "pattern": "^[a-fA-F0-9]+$",
              "description": "Идентификатор сессии или треда: hex 8–64 символов (как `uuid.uuid4().hex`), без дефисов",
              "examples": [
                "a1b2c3d4e5f6789012345678abcdef01"
              ],
              "title": "Thread Id"
            },
            "description": "Идентификатор сессии или треда: hex 8–64 символов (как `uuid.uuid4().hex`), без дефисов"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanReplyBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanReplyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Токен обязателен, но не передан"
          },
          "403": {
            "description": "Неверный Bearer-токен"
          },
          "404": {
            "description": "Тред не найден"
          },
          "422": {
            "description": "Ошибка валидации тела или пути"
          }
        }
      }
    },
...
    "/api/v1/meta/app/{app_id}": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Публичные настройки приложения",
        "description": "Флаги возможностей и инструментов для указанного `app_id` из `config/apps.yaml`. Имя коллекции Chroma и CORS в ответ не попадают.",
        "operationId": "meta_app_public_api_v1_meta_app__app_id__get",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppMetaPublic"
                }
              }
            }
          },
          "404": {
            "description": "Неизвестный app_id"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdminInboxThreadRow": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "app_id": {
            "type": "string",
            "title": "App Id"
          },
          "last_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Preview"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "session_id",
          "app_id"
        ],
        "title": "AdminInboxThreadRow",
        "description": "Строка списка тредов для инбокса оператора."
      },
      "AdminThreadsListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminInboxThreadRow"
            },
            "type": "array",
            "title": "Items"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "limit",
          "offset"
        ],
        "title": "AdminThreadsListResponse"
      },
      "AppMetaPublic": {
        "properties": {
          "app_id": {
            "type": "string",
            "title": "App Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "features": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Features"
          },
          "tools": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Tools"
          },
          "rag_enabled": {
            "type": "boolean",
            "title": "Rag Enabled",
            "default": false
          },
          "ollama_model": {
            "type": "string",
            "title": "Ollama Model",
            "default": ""
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "app_id"
        ],
        "title": "AppMetaPublic",
        "description": "Публичные настройки приложения (без CORS, chroma и прочих внутренних полей)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HumanReplyBody": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 32000,
            "minLength": 1,
            "title": "Content"
          },
          "author_label": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Author Label",
            "description": "Имя или роль, показываемое клиенту (например «Поддержка»)"
          }
        },
        "type": "object",
        "required": [
          "content",
          "author_label"
        ],
        "title": "HumanReplyBody",
        "example": {
          "author_label": "Поддержка",
          "content": "Здравствуйте! Уточните, пожалуйста, последние 4 цифры карты."
        }
      },
      "HumanReplyResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message_id": {
            "type": "integer",
            "title": "Message Id"
          },
          "app_id": {
            "type": "string",
            "title": "App Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "message_id",
          "app_id",
          "session_id",
          "thread_id"
        ],
        "title": "HumanReplyResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Идентификатор приложения из `config/apps.yaml` (например `web_client`, `web_lite`, `web_admin`). Обязателен для чата, RAG-query и `POST /api/v1/chat/completions`. В Swagger: кнопка **Authorize** → API Key.",
        "in": "header",
        "name": "X-App-Id"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Для **`/internal/admin/*`** — значение **`INTERNAL_ADMIN_TOKEN`**; для **`/internal/dev/*`** — **`INTERNAL_DEV_TOKEN`**, если переменная задана на сервере. Иначе заголовок не требуется (только разработка). При **`INTERNAL_AUTH_STRICT`** без настроенного токена — **503**."
      }
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Текущий хост (как в браузере)"
    }
  ]
}